# contracts/openapi/forgeml.v1.openapi.json
{
  "components": {
    "schemas": {
...
      "CreateDatasetRequest": {
        "properties": {
          "description": {
            "default": "",
            "maxLength": 2000,
            "title": "Description",
            "type": "string"
          },
          "name": {
            "maxLength": 120,
            "minLength": 3,
            "title": "Name",
            "type": "string"
          },
          "source_type": {
            "default": "upload",
            "enum": [
              "upload",
              "s3",
              "database",
              "stream"
            ],
            "title": "Source Type",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "CreateDatasetRequest",
        "type": "object"
      },
      "CreateDatasetVersionRequest": {
        "properties": {
          "content_type": {
            "default": "text/csv",
            "maxLength": 255,
            "minLength": 1,
            "title": "Content Type",
            "type": "string"
          },
          "filename": {
            "default": "dataset.csv",
            "maxLength": 255,
            "minLength": 1,
            "title": "Filename",
            "type": "string"
          }
        },
        "title": "CreateDatasetVersionRequest",
        "type": "object"
      },
      "CreateDatasetVersionResponse": {
        "properties": {
          "upload": {
            "$ref": "#/components/schemas/UploadInstructionsResponse"
          },
          "version": {
            "$ref": "#/components/schemas/DatasetVersionResponse"
          }
        },
        "required": [
          "version",
          "upload"
        ],
        "title": "CreateDatasetVersionResponse",
        "type": "object"
      },
...
      "DatasetListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DatasetResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "DatasetListResponse",
        "type": "object"
      },
      "DatasetResponse": {
        "properties": {
          "description": {
            "title": "Description",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "organization_id": {
            "title": "Organization Id",
            "type": "string"
          },
          "project_id": {
            "title": "Project Id",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "source_type": {
            "title": "Source Type",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "organization_id",
          "project_id",
          "name",
          "slug",
          "description",
          "source_type",
          "status"
        ],
        "title": "DatasetResponse",
        "type": "object"
      },
      "DatasetSchemaResponse": {
        "properties": {
          "dataset_version_id": {
            "title": "Dataset Version Id",
            "type": "string"
          },
          "fields": {
            "items": {
              "$ref": "#/components/schemas/SchemaFieldPayload"
            },
            "title": "Fields",
            "type": "array"
          },
          "inferred": {
            "title": "Inferred",
            "type": "boolean"
          },
          "schema_hash": {
            "title": "Schema Hash",
            "type": "string"
          }
        },
        "required": [
          "dataset_version_id",
          "fields",
          "inferred",
          "schema_hash"
        ],
        "title": "DatasetSchemaResponse",
        "type": "object"
      },
      "DatasetValidationRunListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DatasetValidationRunResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "DatasetValidationRunListResponse",
        "type": "object"
      },
      "DatasetValidationRunResponse": {
        "properties": {
          "dataset_version_id": {
            "title": "Dataset Version Id",
            "type": "string"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "report": {
            "additionalProperties": true,
            "title": "Report",
            "type": "object"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "dataset_version_id",
          "status",
          "report",
          "error_message"
        ],
        "title": "DatasetValidationRunResponse",
        "type": "object"
      },
      "DatasetVersionListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DatasetVersionResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "DatasetVersionListResponse",
        "type": "object"
      },
      "DatasetVersionResponse": {
        "properties": {
          "artifact_manifest_hash": {
            "title": "Artifact Manifest Hash",
            "type": "string"
          },
          "artifact_manifest_uri": {
            "title": "Artifact Manifest Uri",
            "type": "string"
          },
          "content_hash": {
            "title": "Content Hash",
            "type": "string"
          },
          "created_by": {
            "title": "Created By",
            "type": "string"
          },
          "dataset_id": {
            "title": "Dataset Id",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "object_uri": {
            "title": "Object Uri",
            "type": "string"
          },
          "row_count": {
            "title": "Row Count",
            "type": "integer"
          },
          "size_bytes": {
            "title": "Size Bytes",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "dataset_id",
          "version",
          "object_uri",
          "content_hash",
          "artifact_manifest_uri",
          "artifact_manifest_hash",
          "row_count",
          "size_bytes",
          "status",
          "created_by"
        ],
        "title": "DatasetVersionResponse",
        "type": "object"
      },
...
      "FinalizeDatasetVersionRequest": {
        "properties": {
          "content_hash": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Content Hash",
            "type": "string"
          },
          "object_uri": {
            "anyOf": [
              {
                "maxLength": 2048,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Object Uri"
          },
          "row_count": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Row Count"
          },
          "sample_csv": {
            "anyOf": [
              {
                "maxLength": 500000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sample Csv"
          },
          "schema_fields": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/SchemaFieldPayload"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schema Fields"
          },
          "size_bytes": {
            "exclusiveMinimum": 0.0,
            "title": "Size Bytes",
            "type": "integer"
          }
        },
        "required": [
          "content_hash",
          "size_bytes"
        ],
        "title": "FinalizeDatasetVersionRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SchemaFieldPayload": {
        "properties": {
          "dtype": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Dtype",
            "type": "string"
          },
          "name": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "nullable": {
            "title": "Nullable",
            "type": "boolean"
          }
        },
        "required": [
          "name",
          "dtype",
          "nullable"
        ],
        "title": "SchemaFieldPayload",
        "type": "object"
      },
...
      "UploadInstructionsResponse": {
        "properties": {
          "expires_at": {
            "title": "Expires At",
            "type": "string"
          },
          "object_uri": {
            "title": "Object Uri",
            "type": "string"
          },
          "required_headers": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Required Headers",
            "type": "object"
          },
          "upload_url": {
            "title": "Upload Url",
            "type": "string"
          }
        },
        "required": [
          "upload_url",
          "object_uri",
          "expires_at",
          "required_headers"
        ],
        "title": "UploadInstructionsResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "ForgeML",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/dataset-versions/{version_id}": {
      "get": {
        "operationId": "get_dataset_version_api_v1_dataset_versions__version_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Version Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetVersionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Dataset Version",
        "tags": [
          "datasets"
        ]
      }
    },
    "/api/v1/dataset-versions/{version_id}/finalize": {
      "post": {
        "operationId": "finalize_dataset_version_api_v1_dataset_versions__version_id__finalize_post",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Version Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FinalizeDatasetVersionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetVersionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Finalize Dataset Version",
        "tags": [
          "datasets"
        ]
      }
    },
    "/api/v1/dataset-versions/{version_id}/schema": {
      "get": {
        "operationId": "get_dataset_schema_api_v1_dataset_versions__version_id__schema_get",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Version Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetSchemaResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Dataset Schema",
        "tags": [
          "datasets"
        ]
      }
    },
    "/api/v1/dataset-versions/{version_id}/validate": {
      "post": {
        "operationId": "validate_dataset_version_api_v1_dataset_versions__version_id__validate_post",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Version Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetValidationRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Validate Dataset Version",
        "tags": [
          "datasets"
        ]
      }
    },
    "/api/v1/dataset-versions/{version_id}/validation-runs": {
      "get": {
        "operationId": "list_dataset_validation_runs_api_v1_dataset_versions__version_id__validation_runs_get",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Version Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetValidationRunListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Dataset Validation Runs",
        "tags": [
          "datasets"
        ]
      }
    },
    "/api/v1/datasets/{dataset_id}": {
      "get": {
        "operationId": "get_dataset_api_v1_datasets__dataset_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "dataset_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Dataset Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Dataset",
        "tags": [
          "datasets"
        ]
      }
    },
    "/api/v1/datasets/{dataset_id}/versions": {
      "get": {
        "operationId": "list_dataset_versions_api_v1_datasets__dataset_id__versions_get",
        "parameters": [
          {
            "in": "path",
            "name": "dataset_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Dataset Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetVersionListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Dataset Versions",
        "tags": [
          "datasets"
        ]
      },
      "post": {
        "operationId": "create_dataset_version_api_v1_datasets__dataset_id__versions_post",
        "parameters": [
          {
            "in": "path",
            "name": "dataset_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Dataset Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDatasetVersionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateDatasetVersionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Dataset Version",
        "tags": [
          "datasets"
        ]
      }
    },
...
    "/api/v1/projects/{project_id}/datasets": {
      "get": {
        "operationId": "list_datasets_api_v1_projects__project_id__datasets_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Datasets",
        "tags": [
          "datasets"
        ]
      },
      "post": {
        "operationId": "create_dataset_api_v1_projects__project_id__datasets_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDatasetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Dataset",
        "tags": [
          "datasets"
        ]
      }
    },
...
  }
}